requirements.txt
pyyaml>=3.12,<4.0
tmuxp>=1.3.3,<2.0.0
